moved prompts lib around
