+++ evals/evals_allower.py
@@ -7,7 +7,7 @@ import time
 from rich.console import Console
 
 from askademic.allower import allower_agent
-from askademic.prompts import USER_PROMPT_ALLOWER_TEMPLATE
+from askademic.prompts.general import USER_PROMPT_ALLOWER_TEMPLATE
 
 
 class AllowerTestCase:
