+++ src/askademic/article.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel, Field
 from pydantic_ai import Agent, Tool
 
 from askademic.constants import GEMINI_2_FLASH_MODEL_ID
-from askademic.prompts import SYSTEM_PROMPT_ARTICLE
+from askademic.prompts.gemini import SYSTEM_PROMPT_ARTICLE
 from askademic.tools import get_article, search_articles_by_title
