+++ src/askademic/main.py
@@ -14,7 +14,7 @@ from rich.console import Console
 from askademic.allower import allower_agent
 from askademic.memory import Memory
 from askademic.orchestrator import orchestrator_agent
-from askademic.prompts import USER_PROMPT_ALLOWER_TEMPLATE
+from askademic.prompts.general import USER_PROMPT_ALLOWER_TEMPLATE
 
 console = Console()
 session = PromptSession(history=FileHistory(".cli_history"))
